hoover3_taskdef::client
Struct Client

pub struct Client {
    pub(crate) inner: ConfiguredClient<TemporalServiceClient<InterceptedService<GrpcMetricSvc, ServiceCallInterceptor>>>,
    pub(crate) namespace: String,
}

Contains an instance of a namespace-bound client for interacting with the Temporal server

Fields
inner: ConfiguredClient<TemporalServiceClient<InterceptedService<GrpcMetricSvc, ServiceCallInterceptor>>>namespace: String
Implementations

impl Client

pub fn new( client: ConfiguredClient<TemporalServiceClient<InterceptedService<GrpcMetricSvc, ServiceCallInterceptor>>>, namespace: String, ) -> Client

Create a new client from an existing configured lower level client and a namespace

pub fn raw_retry_client( &self, ) -> RetryClient<WorkflowServiceClient<InterceptedService<GrpcMetricSvc, ServiceCallInterceptor>>>

Return an auto-retrying version of the underling grpc client (instrumented with metrics collection, if enabled).

Note that it is reasonably cheap to clone the returned type if you need to own it. Such clones will keep re-using the same channel.

pub fn raw_client( &self, ) -> &WorkflowServiceClient<InterceptedService<GrpcMetricSvc, ServiceCallInterceptor>>

Access the underling grpc client. This raw client is not bound to a specific namespace.

Note that it is reasonably cheap to clone the returned type if you need to own it. Such clones will keep re-using the same channel.

pub fn options(&self) -> &ClientOptions

Return the options this client was initialized with

pub fn options_mut(&mut self) -> &mut ClientOptions

Return the options this client was initialized with mutably

pub fn inner( &self, ) -> &ConfiguredClient<TemporalServiceClient<InterceptedService<GrpcMetricSvc, ServiceCallInterceptor>>>

Returns a reference to the underlying client

pub fn into_inner( self, ) -> ConfiguredClient<TemporalServiceClient<InterceptedService<GrpcMetricSvc, ServiceCallInterceptor>>>

Consumes self and returns the underlying client