hoover3_taskdef::tasks
Function init_worker

pub fn init_worker<CT>(
    runtime: &CoreRuntime,
    worker_config: WorkerConfig,
    client: CT,
) -> Result<Worker, Error>
where
    CT: Into<AnyClient>,

Initialize a worker bound to a task queue.

You will need to have already initialized a CoreRuntime which will be used for this worker. After the worker is initialized, you should use CoreRuntime::tokio_handle to run the worker’s async functions.

Lang implementations may pass in a [ConfiguredClient] directly (or a RetryClient wrapping one, or a handful of other variants of the same idea). When they do so, this function will always overwrite the client retry configuration, force the client to use the namespace defined in the worker config, and set the client identity appropriately. IE: Use [ClientOptions::connect_no_namespace], not [ClientOptions::connect].