hoover3_taskdef::tasks
Struct ActContext

pub struct ActContext {
    worker: Arc<dyn Worker>,
    app_data: Arc<AppData>,
    cancellation_token: CancellationToken,
    input: Vec<Payload>,
    heartbeat_details: Vec<Payload>,
    header_fields: HashMap<String, Payload>,
    info: ActivityInfo,
}

Used within activities to get info, heartbeat management etc.

Fields
worker: Arc<dyn Worker>app_data: Arc<AppData>cancellation_token: CancellationTokeninput: Vec<Payload>heartbeat_details: Vec<Payload>header_fields: HashMap<String, Payload>info: ActivityInfo
Implementations

impl ActContext

pub async fn cancelled(&self)

Returns a future the completes if and when the activity this was called inside has been cancelled

pub fn is_cancelled(&self) -> bool

Returns true if this activity has already been cancelled

pub fn extra_inputs(&mut self) -> &mut [Payload]

Retrieve extra parameters to the Activity. The first input is always popped and passed to the Activity function for the currently executing activity. However, if more parameters are passed, perhaps from another languageโs SDK, explicit access is available from extra_inputs

pub fn get_heartbeat_details(&self) -> &[Payload]

Extract heartbeat details from last failed attempt. This is used in combination with retry policy.

pub fn record_heartbeat(&self, details: Vec<Payload>)

RecordHeartbeat sends heartbeat for the currently executing activity

pub fn get_info(&self) -> &ActivityInfo

Get activity info of the executing activity

pub fn headers(&self) -> &HashMap<String, Payload>

Get headers attached to this activity

pub fn app_data<T>(&self) -> Option<&T>
where T: Send + Sync + 'static,

Get custom Application Data