hoover3_taskdef::tasks
Trait TemporalioWorkflowDescriptor

pub trait TemporalioWorkflowDescriptor:
    TemporalioDescriptorName
    + TemporalioDescriptorRegister
    + TemporalioDescriptorValueTypes {
    // Required method
    fn wf_func(
        ctx: WfContext,
        arg: Self::Arg,
    ) -> impl Future<Output = WorkflowResult<Self::Ret>> + Send;

    // Provided methods
    fn register(worker: &mut Worker) -> Result<()> { ... }
    fn workflow_id(arg: &Self::Arg) -> String { ... }
    fn client_start(arg: &Self::Arg) -> impl Future<Output = Result<(), Error>> { ... }
    fn run_as_child(
        wf_ctx: &WfContext,
        arg: Self::Arg,
    ) -> impl Future<Output = Result<Self::Ret, Error>>
       where Self: Sized { ... }
    fn run_parallel(
        wf_ctx: &WfContext,
        args: Vec<Self::Arg>,
    ) -> impl Future<Output = Result<Vec<(Self::Arg, Result<Self::Ret>)>>>
       where Self: Sized { ... }
    fn start_as_child(
        wf_ctx: &WfContext,
        arg: Self::Arg,
    ) -> impl Future<Output = Result<ChildWorkflowFuture<Self>, Error>>
       where Self: Sized { ... }
    fn client_get_status(
        arg: &Self::Arg,
    ) -> impl Future<Output = Result<UiWorkflowStatus, Error>> { ... }
    fn client_get_result(
        arg: &Self::Arg,
    ) -> impl Future<Output = Result<Self::Ret, Error>> { ... }
    fn client_wait_for_completion(
        arg: &Self::Arg,
    ) -> impl Future<Output = Result<Self::Ret, Error>> { ... }
}

Trait implemented by the make_workflow macro.

Required Methods

fn wf_func( ctx: WfContext, arg: Self::Arg, ) -> impl Future<Output = WorkflowResult<Self::Ret>> + Send

The function that implements the workflow.

Provided Methods

fn register(worker: &mut Worker) -> Result<()>

Register the workflow into a given worker

fn workflow_id(arg: &Self::Arg) -> String

Generate a workflow id from an argument, using the stable hash of the argument.

fn client_start(arg: &Self::Arg) -> impl Future<Output = Result<(), Error>>

Start a workflow using a HTTP client. If the workflow already exists, the function returns Ok without restarting it.

fn run_as_child( wf_ctx: &WfContext, arg: Self::Arg, ) -> impl Future<Output = Result<Self::Ret, Error>>
where Self: Sized,

Run the workflow as a child workflow, “blocking” until the child workflow finishes, returning the result of the child workflow.

fn run_parallel( wf_ctx: &WfContext, args: Vec<Self::Arg>, ) -> impl Future<Output = Result<Vec<(Self::Arg, Result<Self::Ret>)>>>
where Self: Sized,

Run the workflow in parallel, returning a vector of results.

fn start_as_child( wf_ctx: &WfContext, arg: Self::Arg, ) -> impl Future<Output = Result<ChildWorkflowFuture<Self>, Error>>
where Self: Sized,

Start the workflow as a child workflow, returning a future that resolves to its result.

fn client_get_status( arg: &Self::Arg, ) -> impl Future<Output = Result<UiWorkflowStatus, Error>>

Get the status of a workflow using a HTTP client. Use this to check if the workflow is running or done or failed.

fn client_get_result( arg: &Self::Arg, ) -> impl Future<Output = Result<Self::Ret, Error>>

Get the result of a workflow using a HTTP client. If the workflow is not done, the function will return errors.

fn client_wait_for_completion( arg: &Self::Arg, ) -> impl Future<Output = Result<Self::Ret, Error>>

Wait for a workflow to complete using HTTP client, sleeping until it finishes, returning the result of the workflow.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.